Add option to view network and disk stats for last 30 seconds
The changes include modifications to HomeController, Client class, and Index View. On the Server side, a boolean flag is added to HomeController's Index action to enable viewing last 30 seconds stats. In the Client model, methods to get network and disk for last 30 seconds are added. In the Index view, the flag is used to display either cumulative or last 30 seconds stats.