Update metrics display and throttle seconds
The metrics.js file was updated to use suggested min and max values instead of set values for scales. This increases flexibility in the chart display. Also, a legend plugin was added but set to false. Furthermore, the throttle time in MetricsController.cs was updated from 0.99 seconds to a full second to improve data capturing.