Update SDK, client, and server versions
The changes in the code involve updating the versions of SDK, client, and server for Aiursoft.StatHub. In addition, some minor changes were made to the wording of comments and the caching pattern in the OsInfoService and HostnameService classes. Package references in the projects are also updated to new versions.